Summary table for unreplicated/fractional 2^k designs.

This function is aimed to provide a summary suitable for unreplicated/fractional designs.

table2kunrep(object)

Arguments

object

fitted model of class aov or lm.

Value

a data.frame containing the estimated effects, the sum of squares (SS), the sum o squares percentual contribution (SSPC), the effects' relative difference with respect to the overal mean (PRD).

Examples

library(planex)
data(semicondutores1rep)
semicondutores1rep <- treatcomb2factors(semicondutores1rep)
fit <- lm(rendimento ~ A*B*C*D*E, data = semicondutores1rep)
table2kunrep(fit)
#>           effects         SS  SSPC    PRD
#> A         11.8125 1116.28125  9.57  38.69
#> B         33.9375 9214.03125 79.00 111.16
#> C          9.6875  750.78125  6.44  31.73
#> D         -0.8125    5.28125  0.05  -2.66
#> E          0.4375    1.53125  0.01   1.43
#> A:B        7.9375  504.03125  4.32  26.00
#> A:C        0.4375    1.53125  0.01   1.43
#> B:C        0.0625    0.03125  0.00   0.20
#> A:D       -0.0625    0.03125  0.00  -0.20
#> B:D       -0.6875    3.78125  0.03  -2.25
#> C:D        0.8125    5.28125  0.05   2.66
#> A:E        0.9375    7.03125  0.06   3.07
#> B:E        0.5625    2.53125  0.02   1.84
#> C:E        0.3125    0.78125  0.01   1.02
#> D:E       -1.1875   11.28125  0.10  -3.89
#> A:B:C     -0.4375    1.53125  0.01  -1.43
#> A:B:D      0.3125    0.78125  0.01   1.02
#> A:C:D     -0.4375    1.53125  0.01  -1.43
#> B:C:D      0.4375    1.53125  0.01   1.43
#> A:B:E     -0.1875    0.28125  0.00  -0.61
#> A:C:E      0.3125    0.78125  0.01   1.02
#> B:C:E      0.9375    7.03125  0.06   3.07
#> A:D:E      0.8125    5.28125  0.05   2.66
#> B:D:E      0.1875    0.28125  0.00   0.61
#> C:D:E     -0.8125    5.28125  0.05  -2.66
#> A:B:C:D   -0.0625    0.03125  0.00  -0.20
#> A:B:C:E    0.1875    0.28125  0.00   0.61
#> A:B:D:E    0.9375    7.03125  0.06   3.07
#> A:C:D:E   -0.3125    0.78125  0.01  -1.02
#> B:C:D:E   -0.9375    7.03125  0.06  -3.07
#> A:B:C:D:E -0.1875    0.28125  0.00  -0.61